+++ .github/actions/rust-build-release/CHANGELOG.md
@@ -3,3 +3,4 @@
 ## [Unreleased]
 
 - Initial skeleton.
+- Replace Bats smoke test with pytest version.